requirements/local.txt
-r base.txt

dal-admin-filters==1.1.0
psycopg2==2.9.3
flake8==4.0.1
black==21.12b0
pre-commit==2.15.0
argh==0.26.2
PyYAML==6.0
watchdog==2.1.6

pytest-django==4.5.2

pytest-factoryboy==2.1.0

Faker==11.3.0
factory-boy==3.2.0

pytest-cov==3.0.0